Add docs > diagrams folder and add Rough State diagram for OTP
Currently, there is no implementation or idea for OTP verification.

Email verification for account creation can be useful to verify emails
(used as usernames).

State diagram is a rough idea to help the above to understand the flow
of how the account verification and emailing works. However, this can be
updated accordingly if there is any need of new changes.