Note: Figured out the issue with the queuing system
Problem lies with asynchronous action for "channel.consume('waiting_users..." where we attempt to locate an user from the waiting queue that chose the same difficulty level as the current user to pair them up together. Due to the asynchronous nature, it does not stop running channel.consume() even if we have we try to return. Hence, the only way to stop it properly is call channel.cancel() with the appropriate tag given for the user in the waiting queue to stop referencing the user PROPERLY in the queue even if IT'S ASYNCHRONOUS.

TO DO:
Eg. Suppose I queue player 1 with email (XX1@gmail.com), difficulty level (easy) then i queue player 2 with email (XX2@gmail.com), difficulty level(easy)
When player 1 is being queued, it will add its message and email as a key pair to the map queueMap, where message is how its being stored in waiting_users in CloudAMQP.
However, when player 2 is being added, i will not be able to retrieve its message from the waiting_users in CloudAMQP as by running channel.consume(), i will get player 1's info.
This allows me to cancel and ACK player 1 but now i need to find out how i get retrieve player 2's message as well so that i can ACK and cancel for player 2 as well!